feat(tools): add read surface for shared work-state — get_work_state tool + resource (#27)
The v0 surface was write-only, so agents had no first-class way to see what was
already claimed before claiming — in dogfooding they cat-ed WORK_STATE.json and
leaned on the CLI. Add a read surface reusing the live status view.

- Relocate buildStatus/renderStatusText/StatusView from cli/commands/status.ts
  to artifacts/work-state-view.ts so both the CLI and tools can use it without
  cli<->tools coupling (status.ts now imports it; behavior unchanged).
- Add a read-only `get_work_state` MCP tool: active claims, overlaps recomputed
  live, and review-ready tasks (text + structured content).
- Expose the same snapshot as an MCP resource `concord://work-state`
  (application/json), which also sets up push notifications (#32).
- End-to-end test via InMemoryTransport confirms both are registered and
  callable through a real MCP client.

Read-only and additive: the three-tool *write* surface is unchanged.
